requirements.txt
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
ekphrasis==0.5.4
ftfy==6.1.3
grpcio==1.60.0
accelerate==0.25.0
transformers==4.32.1
tqdm==4.65.0
ftfy==6.1.3
regex
git+https://github.com/openai/CLIP.git
Pillow==9.4.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorboard==2.15.1